Renamed core/modelmigration to core/migration
