+++ core/migration/phase_internal_test.go
@@ -1,7 +1,7 @@
 // Copyright 2016 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package modelmigration
+package migration
 
 import (
 	"github.com/juju/utils/set"
